feat(crabstar): box futures and use tokio undbounded mpsc for SSR streaming (#1)
Boxing futures in the suspense implementation allows for greater type-safety by allowing to specify concrete generated xxxComplete and should be in general better than the weird futures::streams::select thing I have been using previously.

The coupling with tokio and especially tokio-stream is kinda bad, but seems unavoidable in the long run.
Could replace the concrete tx passed to suspense by impl Sink, but it requires a wrapper type which I do not consider to be worth it.

Also this PR adds a new Page trait for structs that represent individual routes.