feat: remove typed_jinja and typed_jinja_macros
Although the idea of combining askama with minijinja for having the
benefits of both compiled and interpreted template engines is very
enticing, it does not work in the real world.

The differences in syntax make it impossible to use any non-trivial
features, e.g. making reusable "component" templates is very awkward.
On top of it, I think, it should be possible to have reloads without
compilation for separate CSS stylesheets etc. so the slowness of rustc
only affects direct changes to the template, which does not seem to be
a deal-breaker to me.